feat: support for --mkdocs flag
This will allow us a bit more flexibility in consuming mkdocs content. In general, we will need to snarf up the base path from the mkdocs.yml config, because all of the content will be written w.r.t. that base.

This PR also fixes a few issues with the graph and tree formation for "fake" subtasks, i.e. not from imports, but just a way to name a subtask.