fix: use consistent group id for choices
We were using a uuid (v5) for tab group identification. This would not allow for asking the same choice twice. For example: how do you wish to install, ...., how do you wish to uninstall.

The first question is likely to have the same set of choices as the second question. And the same answer...

This changes the tab identifier to be derived in a repeatible fashion from the names of the tabs. It does not yet change any other logic.