feat: pre-validate wizard choice steps
There's no sense in presenting wizard steps that are already done. We were already doing this for task execution, but not for choices. This PR adds that feature.

We will need to find a way to distinguish between choices that are ok not to repeat and choices that bear repeating. This PR assumes all valid choices do not need to be presented again.