chore: refactor tab expansion to occur on Graph model, not AST model
We aren't leveraging this change yet. The goal is to improve support for guidebooks where the outcome of an expansion (e.g. even whether it is successful) depends on the successful execution of a prior task.

For example, you may need to log in to your cloud provider before you can enumerate properties of your account.

To support that kind of complexity will require some changes to the guide logic. It currently assumes that it can resolve all choices before executing any tasks :/