feat: initial support for barrier imports
some guidebooks are going to be “barriers”; a barrier task is one such that choices or task execution coming after it (in the plan) do not make sense until that tasks is executed.

for example, you need to ibmcloud login before you can can offer a choice of resource group

or that you need to target a cluster before you can choose a namespace, etc.

i don’t know how to handle this generally. i think the general solution is to have a special kind of import. if a guidebook “casually” imports another, then this import does not indicate a barrier. e.g. “please have kubectl installed” is an indication of necessity before any tasks can be executed, but may not govern choices… though it may govern some.

perhaps all imports should be indicated as barriers? not sure.

anyway, i have barrier support in place, in the core task graph formulation. how we express and interpret that in the source… we can iterate.