feat: add --veto option, to allow selective override of validation optimization
e.g. to re-do a cloud login, even if you are already logged on

This PR fixes a longer-standing bug in compile.sh, where the tree could be off due to not properly splicing in a new subtree context. The veto support depends on this bug fix, so it seems to make sense to bundle them together, so that tests cover it all.