feat: support "export FOO=bar" and add a "run" task to help test this
e.g.

```shell
madwizard run ./test/inputs/21/in.md
export FOO=3
echo "This should be 3: $FOO"
This should be 3: 3
```

Note: the test for the run task is suboptimal, as it uses a subprocess spawn. This slows things down quite a bit. But fixing this doesn't seem worth saving 15 seconds of test time.