fix: hoistSubTasks optimizer is not order preserving
This PR does not fully fix this problem, but at least gives an ordering that makes more sense. The deeper problem is here:

```typescript
const mySubTasks = removeDuplicates(union(myDominatedSubTasks, frontierAllChoicesSubTasks))
```

this PR gives that swapped order (my, frontier); before it was (frontier, my). The real answer is to find a way to interleave my and frontier subtasks in a way that preserves that initial order of occurence.