feat: initial support for remembered choices
By default now, choices will be stored in a fixed profile (TODO: allow named profiles).

- `--no-profile` CLI option disables profiles altogether
- `--profiles-path` CLI option lets you point to a profiles directory
- `process.env.MWPROFILES_PATH` ibid but via env var

The default profiles path is the platform data directory; e.g. on macOS ~/Library/Application Support/

This PR also does some internal cleanup to avoid Object.assigning Memos and random other stuff. This is helpful, because the default impl `Memoizer` is a class, and so it is important not to mess with `this`. This helps this PR, because we want to pass through extra options, and finally broke things with the way we were smashing memos and options together.

This PR also updates the test rig finally to avoid spawning madwizard as a subprocess for the `run` tests. This helps with this PR, because we will be doing a bunch more run tests, to test the profile support. It is dramatically faster to call madwizard rather than subprocess spawn it (not quite sure why; e.g. 30 seconds down to 6.5 seconds).