fix: "you selected this last time" text shows up in an odd place
we are using enquirer's hint, which places it at the end of the message. but we use newlines in the message, so the hint shows up... oddly.

also, our logic to place a newline between multi-line options is foiled by the sorting we are forced to do, to float the "you selected this last time" option to the top.