feat: support for parsing source into ast at `mirror` time
This is an optimization, where we pre-generate the `blocks` model, and persist it along with the mirror. We do this to work around the extremely slow parsing from the `micromark` markdown parsing node module family.

For a 1.2MB source markdown, this reduces startup time from 7-8 seconds down to nearly 1 second. Node profiling shows that at least 30% of this slowness is due to micromark's array splicing. Oof, we're not going to fix that easily, hence this work.

ref: https://github.com/micromark/micromark/discussions/115