fix: statusMemo is not properly invalidated
The logic was incorrectly invalidating the expansion memo!

This also disables the collapseValidated optimization. It is neither needed nor safe. For example, if we invalidate a `validate` (i.e. statusMemo), then we *also* would need to reoptimize the graph... after every status invalidation. It is far simpler simply not to collapse the graph prematurely. We already check the statusMemo prior to shellExec, anyway, in fe/guide