feat: allow guidebooks to specify a choice via env var
If a tab expansion has the form:
```markdown
=== "expand(expr, message, key)"
```

i.e. it specifies the third argument `key`, then `key` will be interpreted as the name of an environment variable. If that environment variable has a value, then this value will be used not only as the result of the expansion, but also will be forced through as the desired choice.

This feature allows one guidebook to control another. For example, guidebook A may initialize some cluster resources. Later, guidebook B will want to clean up those resources. The names to be cleaned up should match, but guidebook B may also want to be general-purpose to allow any specified (i.e. via  choice) resource to be cleaned up.

See tests/input/41/in.md for an example.