fix: choice descriptions with non-plain text may be cropped
Descriptions that contain links or boldface, or any other non-plain text were being cropped so as only to show the content up to that non-plain content. Though this was only the case for descriptions occuring immediately after the choice marker, without an extra newline.

Was working properly:

```
=== "Choice"

    Description **with bold** and trailing text
```

Was not working, and now fixed by this PR:

```
=== "Choice"
    Description **with bold** and trailing text
```

See tests/inputs/42/in.md for an example.