feat: allow clients to pass in a raw api callback function
Previously, the only option was to interact with the guide via stdin/stdout. Now, you can pass in a raw api handler, if the `raw` option to the cli is a function. This function serves as an `onRaw()` handler that allows for inversion of control (into a UI) via a programmatic interface.

```typescript
export type RawAskEvent = {
  type: "ask"
  ask: Prompt
  onChoose(value: ValidAnswer): void
  onCancel(): void
}

export type RawQADoneEvent = {
  type: "qa-done"
}

export type RawAllDoneEvent = {
  type: "all-done"
  success: boolean
}

export type RawEvent = RawAskEvent | RawQADoneEvent | RawAllDoneEvent
```

BREAKING CHANGE: previously, the `raw` parameter was a boolean, and the "raw prefix" was specified via a separate option. With this PR, `raw` is now either a string (in which case it serves the role of raw prefix); or a function (in which case the alternate callback API is used).