feat: execute finally blocks on abnormal (e.g. ctrl+c) termination
This implements a real runtime finally block stack, so that we only execute finally blocks when the associated block is on stack.

mroe work

Note: this picks up another hack to `enquirer` on the starpit#starpit branch. It is registering an onExit handler that calls process.exit() !! This is from SIGINT. sigh. The extra hack, on commit hash 61450fb17a6f1c9991deaeeb9a4778ac9ffa2382, backs that out. We need to be in charge of when the process exits, if we want finally blocks to work.