feat: switch enquirer from readline to keypress-browserify
enquirer does not need the full power of readline; it appears to be using it (in lib/keypress.js) only to piggyback on the way readline hacks in a keypress event emitter.

NOTE: i had to hack in y18n and escalade into packages/madwizard/package.json. There seems to be a horrible (yes, another) bug in npm... yargs expresses a regular dependence on these two, but npm (both from 16.14 and 16.19) do not install either of these modules... at all, not anywhere.